requirements.txt
requests
python-decouple
SQLAlchemy
numpy
pandas
psycopg2
